api/requirements.txt
# This is a requirements.txt file, which is used to specify the dependencies for a Python project.
# Each line in this file represents a package and its version that the project needs to run successfully.

# The packages are listed in alphabetical order, and their versions are specified after the equal sign.

# The packages are installed using a package manager like pip, which can be invoked from the command line.

# Here are some examples of how to install the packages listed in this file:

# To install all the packages at once, run:
# pip install -r requirements.txt

# To update all the packages to their latest versions, run:
# pip install -U -r requirements.txt

# To install a specific package, run:
# pip install package_name

# To update a specific package, run:
# pip install --upgrade package_name

# The packages listed in this file are essential for the proper functioning of the project.

anyio==4.3.0
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fonttools==4.49.0
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
idna==3.6
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
multitasking==0.0.11
numpy==1.26.4
packaging==23.2
pandas==2.2.1
peewee==3.17.1
pillow==10.2.0
pydantic==1.10.14
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
webencodings==0.5.1
yfinance==0.2.37
